Improved 16b coord handling
Mainly tweaked the code to output models using 16b coords correctly.
They are now output as floating point vals from -1.0 to 1.0.

Minor refactoring of how 16b vertices were created to reduce branching.

Also fixed a bug where if the length of the normal is zero, it still
got output to .obj. Point and Line polygons don't have *one* vertex.
In code, these are created with macroes, and the normals are given as
0,0,0. No normal, no need to output a vn in the .obj.