Cleanup
Cleaned up BufferReader class a little. Removed several unused and de-facto non-implemented functions from the class.

Added braces to guard a couple of statements. Without them, the second statement on the line is executed whether or not it needs to be. If it doesn't need to be, it doesn't hurt anything to execute. Cleared a -Wmisleading-intentation warning.